Update tests for z/OS compatiability
Introduce minor test changes to improve compatibility with z/OS.

- Add new config.testConnectionStrings.zos.json and code in common.js
to support z/OS ODBC connectionString format (DSN, UID, PWD).
- Introduce IBM_DB_SERVER_TYPE env var to guard code specific for
connecting to z/OS.
- Update NUMERIC type to NUM.
- Change blob file paths to absolute path.
- Update expected Error Code and Strings for z/OS ODBC
- Separate multi-row inserts and combined selects (not supported for z/OS)
- Capitalize specific keywords for szTableType
- Capitalize procedure names.
- Change CREATE OR REPLACE to DROP then CREATE (z/OS only).
- Add "disable debug mode" to bypass WLM requirement.
- Add explicit CCSID to UNICODE to ensure consistent string constant encodings

+++ test/test-sp-resultset-execute.js
@@ -9,11 +9,29 @@ var common = require("./common")
 
 if(schema == undefined) schema = "NEWTON";
    
-var proc1 = "create or replace procedure " + schema + ".proc2 ( IN v1 int, INOUT v2 varchar(30) )  dynamic result sets 2 language sql begin  declare cr1  cursor with return for select c1, c2 from " + schema + ".mytab1; declare cr2  cursor with return for select c2 from " + schema + ".mytab1; open cr1; open cr2; set v2 = 'success'; end";
-var proc2 = "create or replace procedure " + schema + ".proc2 ( IN v1 int, INOUT v2 varchar(30) )  language sql begin  set v2 = 'success'; end";
-var proc3 = "create or replace procedure " + schema + ".proc2 ( IN v1 int, IN v2 varchar(30) )  dynamic result sets 2 language sql begin  declare cr1  cursor with return for select c1, c2 from " + schema + ".mytab1; declare cr2  cursor with return for select c2 from " + schema + ".mytab1; open cr1; open cr2; end";
-var query = "call " + schema + ".proc2(?, ?)";
+var proc1 = "create or replace procedure " + schema + ".PROC2 ( IN v1 int, INOUT v2 varchar(30) )  dynamic result sets 2 language sql begin  declare cr1  cursor with return for select c1, c2 from " + schema + ".mytab1; declare cr2  cursor with return for select c2 from " + schema + ".mytab1; open cr1; open cr2; set v2 = 'success'; end";
+var proc2 = "create or replace procedure " + schema + ".PROC2 ( IN v1 int, INOUT v2 varchar(30) )  language sql begin  set v2 = 'success'; end";
+var proc3 = "create or replace procedure " + schema + ".PROC2 ( IN v1 int, IN v2 varchar(30) )  dynamic result sets 2 language sql begin  declare cr1  cursor with return for select c1, c2 from " + schema + ".mytab1; declare cr2  cursor with return for select c2 from " + schema + ".mytab1; open cr1; open cr2; end";
+var query = "call " + schema + ".PROC2(?, ?)";
 var result;
+
+if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+  // Update the queries for Db2 on z/OS compatiability.
+  // - Does not support CREATE OR REPLACE syntax.
+ var dropProc = "drop procedure " + schema + ".PROC2";
+  proc1 = proc1.replace(" or replace", "");
+  proc2 = proc2.replace(" or replace", "");
+  proc3 = proc3.replace(" or replace", "");
+
+  // - When creating an SQL native procedure on z/OS, you will need to have a WLM environment
+  // defined for your system if you want to run the procedure in debugging mode. Adding
+  // "disable debug mode" to bypass this requirement.
+  proc1 = proc1.replace(" begin", " disable debug mode begin");
+  proc2 = proc2.replace(" begin", " disable debug mode begin");
+  proc3 = proc3.replace(" begin", " disable debug mode begin");
+}
+
+
 //ibmdb.debug(true);
 ibmdb.open(common.connectionString, {fetchMode : 3}, function (err, conn) {
     if(err) { 
@@ -24,10 +42,23 @@ ibmdb.open(common.connectionString, {fetchMode : 3}, function (err, conn) {
       conn.querySync("drop table " + schema + ".mytab1");
     } catch (e) {};
     conn.querySync({"sql":"create table " + schema + ".mytab1 (c1 int, c2 varchar(20))", "noResults":true});
-    conn.querySync("insert into " + schema + ".mytab1 values (2, 'bimal'), (3, 'kumar')");
+    if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+      // Db2 on z/OS does not support multi-row inserts
+      conn.querySync("insert into " + schema + ".mytab1 values (2, 'bimal')");
+      conn.querySync("insert into " + schema + ".mytab1 values (3, 'kumar')");
+    } else {
+      conn.querySync("insert into " + schema + ".mytab1 values (2, 'bimal'), (3, 'kumar')");
+    }
+
     param2 = {ParamType:"INOUT", DataType:1, Data:"abc", Length:50};
 
     // Create SP with INOUT param and 2 Result Set.
+    if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+      // CREATE OR REPLACE is not supported on z/OS.  Explicitly drop procedure.
+      try {
+         conn.querySync(dropProc);
+      } catch(e) { };
+    }
     conn.querySync(proc1);
     // Call SP Synchronously.
     stmt = conn.prepareSync(query);
@@ -71,6 +102,12 @@ ibmdb.open(common.connectionString, {fetchMode : 3}, function (err, conn) {
         stmt.closeSync();
 
         // Create SP with only OUT param and no resultset.
+        if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+          // CREATE OR REPLACE is not supported on z/OS.  Explicitly drop procedure.
+          try {
+            conn.querySync(dropProc);
+          } catch(e) { };
+        }
         conn.querySync(proc2);
         // Call SP Synchronously.
         stmt = conn.prepareSync(query);
@@ -101,6 +138,12 @@ ibmdb.open(common.connectionString, {fetchMode : 3}, function (err, conn) {
             stmt.closeSync();
 
             // Create SP with only Result Set and no OUT or INPUT param.
+            if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+              // CREATE OR REPLACE is not supported on z/OS.  Explicitly drop procedure.
+              try {
+                conn.querySync(dropProc);
+              } catch(e) { };
+            }
             conn.querySync(proc3);
             // Call SP Synchronously.
             stmt = conn.prepareSync(query);
@@ -141,7 +184,11 @@ ibmdb.open(common.connectionString, {fetchMode : 3}, function (err, conn) {
                 stmt.closeSync();
 
                 // Do Cleanup.
-                conn.querySync("drop procedure " + schema + ".proc2 ( INT, VARCHAR(30) )");
+		if (process.env.IBM_DB_SERVER_TYPE === "ZOS") {
+                  conn.querySync("drop procedure " + schema + ".PROC2");
+		} else {
+                  conn.querySync("drop procedure " + schema + ".PROC2 ( INT, VARCHAR(30) )");
+                }
                 conn.querySync("drop table " + schema + ".mytab1");
                 // Close connection in last only.
                 conn.close(function (err) { console.log("done.");});
